הסרה והעברה לפונקצייה חדשה
