+++ python-project/pro-func/Folder_Merger.py
@@ -11,7 +11,11 @@
 8. לבצע מיזוג תיקיות לפי הרשימה.
 '''
 
-import csv, os, sys, shutil, identify_similarities
+import csv, os, sys, shutil
+# יבוא פונקציה לזיהוי דמיון בין מחרוזות
+from identify_similarities import similarity_sure
+# פונקציית ניקוי מסך
+from click import clear
 
 # יבוא רשימת זמרים מקובץ csv
 def read_csv(file_path):
@@ -41,6 +45,9 @@ def merge_folders(singers_data, dir_path):
         # אם שם תיקית המקור והיעד זהים תתבצע חזרה להמשך הלולאה
         if source_name == target_name: continue
         
+        # בדיקה אם שם האמן קיים ברשימת התיקיות
+        if not source_name in os.listdir(): continue
+        
         # הגדרת נתיבי תיקית מקור ותיקית יעד
         old_path = os.path.join(os.getcwd(), source_name)
         new_path = os.path.join(os.getcwd(), target_name)
@@ -51,7 +58,7 @@ def merge_folders(singers_data, dir_path):
             print(old_path + ' --> \n' + new_path)
        
         # אם קיים נתיב יעד, תתבצע העברה של הקבצים שבמקור אל היעד
-        elif os.path.exists(old_path):       
+        elif os.path.exists(old_path):
             for filename in os.listdir(old_path):
                 source_path = os.path.join(old_path, filename)
                 destination_path = os.path.join(new_path, filename)
@@ -61,11 +68,99 @@ def merge_folders(singers_data, dir_path):
             shutil.rmtree(old_path)
 
 
+# בדיקה אם שם האמן קיים כבר בצורה דומה
+def check_similarity(target_dir, artist):
+    """
+בדיקה אם שם אמן קיים ברשימת תיקיות
+
+פרמטרים:
+    פרמטר 1 = נתיב תיקיה
+    פרמטר 2 = שם אמן
+
+תוצאה:
+    שם האמן הדומה או "None"
+    """
+    list_dirs = os.listdir(target_dir)
+    # יציאה מהפונקציה במקרה ורשימת הקבצים ריקה
+    if list_dirs == []:
+        return None
+    # בדיקת דמיון בין מחרוזות כדי לבדוק אם קיים שם אמן דומה בתיקית היעד
+    answer, similarity_str = similarity_sure(artist, list_dirs, False)
+    if answer:
+        return similarity_str
+    else:
+        return None
+
+
+#יצירת רשימת שמות דומים לפי בחירת המשתמש
+def similarity_list(dir_path):
+
+    # הגדרת סט שמות אמנים דומים    
+    similarity_set = set()
+    not_similarity_set = set()
+
+    for artist in os.listdir(dir_path):
+        # הפעלת בדיקה אם שם אמן דומה כבר קיים ביעד
+        similarity_str = check_similarity(dir_path, artist)
+        set_item = (artist, similarity_str)
+          
+        # בדיקה אם השם כבר מופיע ברשימת הדומים / הלא דומים
+        if similarity_str and not set_item in similarity_set \
+            and not set_item in not_similarity_set:
+            # מתן אפשרות למשתמש לבחור אם למזג את שמות הזמרים
+            print('{}\n"{}" {} "{}"\n{}'.format("נמצאו שמות דומים - למזג?", artist, "<-->", similarity_str, "הקש 1 לאישור או 2 להמשך"))
+            answer = input(">>>")
+            # ניקוי מסך
+            clear()
+            try:
+                if int(answer) == 1:
+                    similarity_set.add(set_item)
+                    artist = similarity_str
+                elif int(answer) == 2:
+                    not_similarity_set.add(set_item)
+            except:
+                pass
+
+        elif (artist, similarity_str) in similarity_set:
+            artist = similarity_str
+    
+    # מיזוג זוגות שמתו דומים לקבוצות
+    merges_list = merge_tuples(similarity_set)
+    print(merges_list)
+
+
+# מיזוג טאפלים המכילים שמות זמרים זהים
+def merge_tuples(input_set):
+    '''
+ממזג טאפלים המכילים מחרוזת אחת זהה לרשימה חדשה המכילה טאפלים
+    פרמטר = סט המכיל טאפלים בני 2 מחרוזות
+    
+    תוצאה = רשימה עם סטים ממוזגים    
+    '''
+    result = list(input_set)
+    i = 0
+    while i < len(result):
+        v = result[i]
+        for j, tup in enumerate(result):
+            if v == tup:
+                continue
+            if not set(v).isdisjoint(set(tup)):
+                result[j] = set(v).union(set(tup))
+                result.pop(i)
+                break
+        else:
+            i += 1
+
+    return result
+
+
+
 def main():
     dir_path = str(sys.argv[1])
     file_path = r"C:\Users\COLMI\AppData\Roaming\singles-sorter\singer-list.csv"
     singers_data = read_csv(file_path)
     merge_folders(singers_data, dir_path)
+    similarity_list(dir_path)
 
 if __name__ == '__main__':
     main()
